feat(parity): aggregates + time-ops + topops + box parity; fix tstz distance
Parity tests activated (21/23 files now pass; 2 tile/split files remain
skipped pending table-function architecture, tracked by PRs #102/#110):

  015_span_aggfuncs   — SetUnionAgg / SpanUnionAgg / extent (20 assertions)
  022_temporal        — temporal accessors + operators (32 assertions)
  032_temporal_box    — tbox constructors + expandValue/expandTime (3 assertions)
  032_temporal_topops — temporal @> tbox / && temporal / -|- tstzspan (3 assertions)
  036_tnumber_distance — tint<->tint, nad, nearestApproachDistance (3 assertions)
  040_temporal_aggfuncs — TcountAgg/TandAgg/TminAgg/TsumAgg/extent (5 assertions)
  042_temporal_waggfuncs — WminAgg/WmaxAgg/WsumAgg (3 assertions)
  050_geo_inventory   — geo surface manifest activated (1 assertion)
  009_time_ops        — tstzset/tstzspan + interval shift; tstzspan <-> INTERVAL (4 assertions)

Bug fixes:
  - tstzspan/tstzset <-> operator: was registered as INTERVAL but implementation
    returned double, causing INTERNAL crash. Fixed Distance_span_value,
    Distance_value_span (use distance_span_timestamptz), Distance_span_span
    (dispatch on T_TSTZSPAN), Distance_set_value/Distance_value_set/
    Distance_set_set (tstzset case). All time-type distances now return
    interval_t normalised to days+micros so "2 days" formats correctly.
  - expandValue(tbox, INTEGER): TboxExpandValueExecutor<int64_t> caused
    INTERNAL error "Expected vector of type INT64, but found vector of type
    INT32". Fixed template parameter to int32_t.
  - SetUnionAgg(date): date_to_set called raw DuckDB date_t without epoch
    offset conversion, producing dates 30 years in future. Added
    date_to_set_duckdb wrapper using ToMeosDate().

New registrations:
  - "+", {tstzset, INTERVAL} → tstzset: shift alias (MobilityDB @> parity)
  - "+", {tstzspan, INTERVAL} → tstzspan: shift alias
  - SecondsToInterval helper in time_util.hpp

+++ src/temporal/tbox_functions.cpp
@@ -1049,7 +1049,7 @@ void TboxFunctions::TboxExpandValueExecutor(Vector &tbox, Vector &value, meosTyp
 void TboxFunctions::Tbox_expand_value(DataChunk &args, ExpressionState &state, Vector &result) {
     const auto &arg_type = args.data[1].GetType();
     if (arg_type.id() == LogicalTypeId::INTEGER) {
-        TboxExpandValueExecutor<int64_t>(args.data[0], args.data[1], T_INT4, result, args.size());
+        TboxExpandValueExecutor<int32_t>(args.data[0], args.data[1], T_INT4, result, args.size());
     } else if (arg_type.id() == LogicalTypeId::DOUBLE) {
         TboxExpandValueExecutor<double>(args.data[0], args.data[1], T_FLOAT8, result, args.size());
     } else {
